requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
python-jose
requests
psycopg2-binary
pydantic[email]
email-validator
firebase-admin
cloudinary
redis[async]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic
reportlab
apscheduler
resend
openai
google-genai
python-multipart